Move commands, keeping tests in place
